requirements.txt
bertopic==0.16.0
blinker==1.7.0
cachelib==0.12.0
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Cython==0.29.37
filelock==3.13.1
Flask==3.0.2
Flask-Session==0.6.0
fsspec==2024.2.0
google-api-core==2.17.1
google-api-python-client==2.119.0
google-auth==2.28.1
google-auth-httplib2==0.2.0
googleapis-common-protos==1.62.0
gunicorn==21.2.0
hdbscan==0.8.33
httplib2==0.22.0
huggingface-hub==0.20.3
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
llvmlite==0.42.0
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numba==0.59.0
numpy==1.26.4
packaging==23.2
pandas==2.2.1
pillow==10.2.0
pipdeptree==2.16.1
plotly==5.19.0
protobuf==4.25.3
psutil==5.9.8
pyasn1==0.5.1
pyasn1-modules==0.3.0
pynndescent==0.5.11
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
redis==5.0.3
regex==2023.12.25
requests==2.31.0
rsa==4.9
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
sentence-transformers==2.4.0
six==1.16.0
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.3.0
tokenizers==0.15.2
torch==2.2.1
tqdm==4.66.2
transformers==4.38.1
typing_extensions==4.10.0
tzdata==2024.1
umap-learn==0.5.5
uritemplate==4.1.1
urllib3==2.2.1
Werkzeug==3.0.1
